Move window fixture to conftest.py, create test file for tile_calculator.py
